Update of http_version.pl
This is because otherwise a different commit is save to localStorage in version.js when saveCheckpoint is called the first time and when clicking on new. In the first case, changes is asked without parameters, in the latter, with the stored commit. If the last commit is not tagged, in the first case it is returned, while in latter case the latest tagged commit is returned. This causes a wrong recognition of changes